pyproject.toml
[tool.poetry]
name = "note-it"
version = "0.1.0"
description = ""
authors = ["makefinks <leon.fink.lf@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
pymupdf = "^1.23.26"
streamlit = "^1.32.0"
anthropic = "^0.19.1"
markdown2 = "^2.4.13"
pdfkit = "^1.0.0"
pypdf2 = "^3.0.1"
wkhtmltopdf = "^0.2"
opencv-python = "^4.9.0.80"
pillow = "^10.2.0"
streamlit-code-editor = "^0.1.12"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
